+++ src/converter.cpp
@@ -3,6 +3,7 @@
 #include <cstdio>
 #include <cstring>
 #include <ctime>
+#include <filesystem>
 #include <functional>
 #include <iomanip>
 #include <memory>
@@ -43,6 +44,35 @@ void Converter::set_id(const uint64_t _id) { id = _id; }
  * PacketConverter
  */
 PacketConverter::PacketConverter(const uint32_t _l2_type) : l2_type(_l2_type) {}
+
+PacketConverter::PacketConverter(const uint32_t _l2_type,
+                                 const time_t launch_time)
+    : l2_type(_l2_type)
+{
+  if (!session_file.is_open()) {
+    struct tm* ts;
+    ts = localtime(&launch_time);
+    char buf[80];
+
+    if (std::filesystem::is_directory("/report")) {
+      strftime(buf, sizeof(buf), "/report/sessions.txt-%Y%m%d%H%M%S", ts);
+    } else {
+      strftime(buf, sizeof(buf), "sessions.txt-%Y%m%d%H%M%S", ts);
+    }
+    session_file.open(buf, ios::out | ios::app);
+  }
+}
+
+void PacketConverter::save_session(uint64_t event_id, uint32_t src,
+                                   uint32_t dst, uint8_t proto, uint16_t sport,
+                                   uint16_t dport)
+{
+  if (session_file.is_open()) {
+    session_file << event_id << "," << src << "," << dst << ","
+                 << (static_cast<uint16_t>(proto)) << "," << sport << ","
+                 << dport << endl;
+  }
+}
 
 Conv::Status PacketConverter::convert(uint64_t event_id, char* in,
                                       size_t in_len, PackMsg& pm)
@@ -288,7 +318,7 @@ void PacketConverter::update_pack_message(uint64_t event_id, PackMsg& pm,
                                           const char* in, size_t in_len)
 {
   if (in == nullptr && in_len == 0) {
-    id = sessions.make_next_message(pm, id);
+    sessions.make_next_message(pm, event_id);
     return;
   }
   if (in == nullptr || in_len == 0) {
@@ -296,13 +326,16 @@ void PacketConverter::update_pack_message(uint64_t event_id, PackMsg& pm,
   }
   size_t data_offset =
       sizeof(pcap_sf_pkthdr) + sizeof(ether_header) + ip_hl + l4_hl + vlan;
-  sessions.update_session(event_id, src, dst, proto, sport, dport,
-                          in + data_offset, in_len - data_offset);
+  if (sessions.update_session(src, dst, proto, sport, dport, in + data_offset,
+                              in_len - data_offset)) {
+    save_session(event_id, src, dst, proto, sport, dport);
+  }
+
   size_t estimated_data =
       (sessions.size() * (session_extra_bytes + message_n_label_bytes)) +
       sessions.get_number_bytes_in_sessions() + pm.get_bytes();
   if (estimated_data >= pm.get_max_bytes()) {
-    id = sessions.make_next_message(pm, id);
+    sessions.make_next_message(pm, event_id);
   }
 }